pyproject.toml
[project]
name = "origami-jsynth"
version = "0.1.0"
description = "Minimal reproduction package for Origami tabular/JSON synthesis experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "origami-ml==2.0.0a10",
    "pandas>=2.3,<3",
    "scikit-learn>=1.8",
    "scipy>=1.14",
    "xgboost>=3.1",
    "numba>=0.63",
    "huggingface-hub>=0.20",
    "pyyaml>=6.0",
]

[project.scripts]
origami-jsynth = "origami_jsynth.cli:main"

[dependency-groups]
dev = ["ruff>=0.9"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["SIM105"]
